broke up GUIDTask tests into separate files; attempting to increase code coverage by modifying the first registering test
