+++ common/src/test/scala/objects/guidtask/GUIDTaskUnregister1Test.scala
@@ -1,0 +1,18 @@
+// Copyright (c) 2017 PSForever
+package objects.guidtask
+
+import net.psforever.objects.guid.{GUIDTask, TaskResolver}
+import objects.ActorTest
+
+class GUIDTaskUnregister1Test extends ActorTest() {
+  "UnregisterObjectTask" in {
+    val (guid, uns, taskResolver, probe) = GUIDTaskTest.CommonTestSetup
+    val obj = new GUIDTaskTest.TestObject
+    guid.register(obj, "dynamic")
+
+    assert(obj.HasGUID)
+    taskResolver ! TaskResolver.GiveTask(new GUIDTaskTest.RegisterTestTask(probe.ref), List(GUIDTask.UnregisterObjectTask(obj)(uns)))
+    probe.expectMsg(scala.util.Success)
+    assert(!obj.HasGUID)
+  }
+}
